re arranged files and functions
